Remove Wasm bindgen from config types, move render config into a submodule
